+++ src/nemseer/__init__.py
@@ -4,9 +4,11 @@ import sys
 from importlib.metadata import version
 
 # aliases for SQLLoader functions
+from .data import FORECAST_TYPES as forecast_types
 from .downloader import get_sqlloader_forecast_tables as get_tables
 from .downloader import get_sqlloader_years_and_months as get_data_daterange
-from .nemseer import download_raw_data
+from .forecast_type.run_time_generators import generate_runtimes
+from .nemseer import compile_raw_data, download_raw_data
 
 __version__ = version("nemseer")
 
@@ -15,7 +17,11 @@ logging.basicConfig(
     stream=sys.stdout, level=logging.INFO, format="%(levelname)s: %(message)s"
 )
 
-__all__ = ["download_raw_data", "get_tables", "get_data_daterange"]
-
-forecast_types = ("P5MIN", "PREDISPATCH", "PDPASA", "STPASA", "MTPASA")
-"""Forecast types requestable through nemseer"""
+__all__ = [
+    "compile_raw_data",
+    "download_raw_data",
+    "get_tables",
+    "get_data_daterange",
+    "generate_runtimes",
+    "forecast_types",
+]