+++ src/nemseer/downloader.py
@@ -13,7 +13,14 @@ from attrs import define, field
 from bs4 import BeautifulSoup
 from tqdm.auto import tqdm
 
-from .data import ENUMERATED_TABLES, MMSDM_ARCHIVE_URL, PREDISP_ALL_DATA, USER_AGENTS
+from .data import (
+    ENUMERATED_TABLES,
+    FORECAST_TYPES,
+    INVALID_STUBS_FILE,
+    MMSDM_ARCHIVE_URL,
+    PREDISP_ALL_DATA,
+    USER_AGENTS,
+)
 from .data_handlers import clean_forecast_csv
 from .query import (
     Query,
@@ -27,9 +34,8 @@ logger = logging.getLogger(__name__)
 
 def _validate_forecast_type(forecast_type: str):
     """Check user-supplied forecast type is valid"""
-    valid_types = ("P5MIN", "PREDISPATCH", "PDPASA", "STPASA", "MTPASA")
-    if forecast_type not in valid_types:
-        raise ValueError(f"Forecast type should be one of {valid_types}")
+    if forecast_type not in FORECAST_TYPES:
+        raise ValueError(f"Forecast type should be one of {FORECAST_TYPES}")
 
 
 def _build_useragent_generator(n: int) -> Generator[str, None, None]:
@@ -203,9 +209,10 @@ def get_sqlloader_forecast_tables(
 ) -> List[str]:
     """Requestable tables of particular forecast type on MMSDM Historical Data SQLLoader
 
-    If `actual` = False, provides a list of tables that can be requested via `nemseer`.
+    If :attr:`actual` = False, provides a list of tables that can be requested via
+    `nemseer`.
 
-    If `actual` = True, returns actual tables available via NEMWeb, including all
+    If :attr:`actual` = True, returns actual tables available via NEMWeb, including all
     tables that are enumerated.
 
     N.B.:
@@ -288,7 +295,9 @@ def get_sqlloader_years_and_months() -> Dict[int, List[int]]:
 
 
 def get_unzipped_csv(url: str, raw_cache: Path) -> None:
-    """Unzipped (single) csv file downloaded from `url` to :attr:`raw_cache`
+    """Unzipped (single) csv file downloaded from `url` to :term:`raw_cache`
+
+    This function:
 
     1. Downloads zip file in chunks to limit memory use and enable progress bar
     2. Validates that the zip contains a single file that has the same name as the zip
@@ -296,7 +305,7 @@ def get_unzipped_csv(url: str, raw_cache: Path) -> None:
 
     Args:
         url: URL of zip
-        raw_cache: Path to save zip
+        raw_cache: Path to save zip. See :term:`raw_cache`.
     Returns:
         None. Extracts csvs to :attr:`raw_cache`.
     """
@@ -333,7 +342,7 @@ def get_unzipped_csv(url: str, raw_cache: Path) -> None:
             z.close()
         except BadZipFile:
             logging.error(f"{z.testzip()} invalid or corrupted")
-            invalid_files = raw_cache / Path(".invalid_aemo_files.txt")
+            invalid_files = raw_cache / Path(INVALID_STUBS_FILE)
             _invalid_zip_to_file(invalid_files, fn.group(1))
         Path(file_path).unlink()
     else:
@@ -363,6 +372,19 @@ def _validate_tables_on_run_start(instance, attribute, value) -> None:
 
 @define(kw_only=True)
 class ForecastTypeDownloader:
+    """:class:`ForecastTypeDownloader` can initiate csv downloads and convert
+    :term:`raw_cache` csvs to the parquet format.
+
+    Attributes:
+        run_start: Forecast runs at or after this datetime are queried.
+        run_end: Forecast runs before or at this datetime are queried.
+        forecast_type: One of :data:`nemseer.forecast_types`
+        tables: Table or tables required. A single table can be supplied as
+            a string. Multiple tables can be supplied as a list of strings.
+        raw_cache: Path to download raw data to. Can reuse or build a
+            new :term:`raw_cache`.
+    """
+
     run_start: datetime
     run_end: datetime
     forecast_type: str
@@ -370,7 +392,7 @@ class ForecastTypeDownloader:
     raw_cache: Path
 
     @classmethod
-    def from_Query(cls, query: Query):
+    def from_Query(cls, query: Query) -> "ForecastTypeDownloader":
         """Constructor method for :class:`ForecastTypeDownloader` from
         :class:`Query <nemseer.query.Query>`"""
         tables = query.tables
@@ -398,7 +420,7 @@ class ForecastTypeDownloader:
         filename_data = generate_sqlloader_filenames(
             self.run_start, self.run_end, self.forecast_type, self.tables
         )
-        invalid_or_corrupted_stubfile = self.raw_cache / ".invalid_aemo_files.txt"
+        invalid_or_corrupted_stubfile = self.raw_cache / Path(INVALID_STUBS_FILE)
         for metadata in filename_data.keys():
             fname = filename_data[metadata]
             (year, month, table) = metadata
@@ -428,8 +450,10 @@ class ForecastTypeDownloader:
     def convert_to_parquet(self, keep_csv=False) -> None:
         """Converts all CSVs in the :attr:`raw_cache` to parquet
 
-        Logs a warning if the filesize is greater than half of available memory as
-        :class:`pandas.DataFrame` consumes more than the file size in memory.
+        Warning:
+            A warning is printed if the filesize is greater than half of available
+            memory as :class:`pandas.DataFrame` consumes more than the file size in
+            memory.
         """
         csvs = list(Path(self.raw_cache).glob("*.[Cc][Ss][Vv]"))
         for csv in csvs: