+++ tests/test_nemseer.py
@@ -1,26 +1,29 @@
 import logging
+from datetime import datetime, timedelta
 
-import grequests  # type: ignore
+import pandas as pd
 import pytest
-import requests
 
-from nemseer import download_raw_data, forecast_types, get_tables
-from nemseer.downloader import (
-    _build_useragent_generator,
-    _construct_sqlloader_forecastdata_url,
+from nemseer import compile_raw_data, download_raw_data
+from nemseer.data import (
+    DATETIME_FORMAT,
+    FORECASTED_COL,
+    INVALID_STUBS_FILE,
+    RUNTIME_COL,
 )
+from nemseer.forecast_type.run_time_generators import generate_runtimes
+from nemseer.query import generate_sqlloader_filenames
 
 
 class TestDowloadRawData:
     def test_download_and_query_check(self, caplog, download_file_to_cache):
         query = download_file_to_cache
         caplog.set_level(logging.INFO)
-        date_strformat = "%Y/%m/%d %H:%M"
         download_raw_data(
-            query.run_start.strftime(date_strformat),
-            query.run_end.strftime(date_strformat),
-            query.forecasted_start.strftime(date_strformat),
-            query.forecasted_end.strftime(date_strformat),
+            query.run_start.strftime(DATETIME_FORMAT),
+            query.run_end.strftime(DATETIME_FORMAT),
+            query.forecasted_start.strftime(DATETIME_FORMAT),
+            query.forecasted_end.strftime(DATETIME_FORMAT),
             query.forecast_type,
             query.tables,
             query.raw_cache,
@@ -34,19 +37,163 @@ class TestDowloadRawData:
         )
 
 
-@pytest.mark.parametrize("ftype", forecast_types)
-class TestAllTableRequests:
-    def test_all_table_requests_valid(self, ftype, get_test_year_and_month):
-        def _check_size(response: requests.Response):
-            size = int(response.headers.get("Content-Length", 0))
-            assert size > 100
+class TestCompileRawData:
+    def setup_compilation_test(
+        self, gen_datetime, fix_forecasted_dt, forecast_type, time_delta
+    ):
+        forecasted_start = gen_datetime
+        forecasted_start = fix_forecasted_dt(forecasted_start, forecast_type)
+        forecasted_end = forecasted_start + time_delta
+        forecasted_start = forecasted_start.strftime(DATETIME_FORMAT)
+        forecasted_end = forecasted_end.strftime(DATETIME_FORMAT)
 
-        year, month = get_test_year_and_month
-        ftype_tables = get_tables(year, month, ftype)
-        useragents = _build_useragent_generator(len(ftype_tables))
-        reqs = []
-        for table in ftype_tables:
-            url = _construct_sqlloader_forecastdata_url(year, month, ftype, table)
-            reqs.append(grequests.get(url, headers={"User-Agent": next(useragents)}))
-        for resp in grequests.imap(reqs, size=len(reqs)):
-            _check_size(resp)
+        (str_start, str_end) = (forecasted_start, forecasted_end)
+        run_start, run_end = generate_runtimes(str_start, str_end, forecast_type)
+        return run_start, run_end, forecasted_start, forecasted_end
+
+    def test_invalid_files_in_query(
+        self,
+        gen_datetime,
+        fix_forecasted_dt,
+        mocker,
+        tmp_path,
+        caplog,
+    ):
+        def mock_pd_concat(dfs, axis=0):
+            pass
+
+        (forecast_type, table) = ("STPASA", "REGIONSOLUTION")
+        time_delta = timedelta(hours=72)
+        (
+            run_start,
+            run_end,
+            forecasted_start,
+            forecasted_end,
+        ) = self.setup_compilation_test(
+            gen_datetime, fix_forecasted_dt, forecast_type, time_delta
+        )
+        fnames = generate_sqlloader_filenames(
+            datetime.strptime(run_start, DATETIME_FORMAT),
+            datetime.strptime(run_end, DATETIME_FORMAT),
+            forecast_type,
+            [table],
+        ).values()
+        stubfile = tmp_path / INVALID_STUBS_FILE
+        with open(stubfile, "x") as f:
+            for fn in fnames:
+                f.write(f"{fn}\n")
+        caplog.set_level(logging.WARNING)
+        mocker.patch("nemseer.data_compilers.pd.concat", mock_pd_concat)
+        compile_raw_data(
+            run_start,
+            run_end,
+            forecasted_start,
+            forecasted_end,
+            forecast_type,
+            table,
+            tmp_path,
+        )
+        assert all(
+            [INVALID_STUBS_FILE in record.msg for record in caplog.get_records("call")]
+        )
+
+    def test_invalid_format(
+        self,
+        gen_datetime,
+        fix_forecasted_dt,
+        tmp_path,
+    ):
+        (forecast_type, table) = ("STPASA", "INTERCONNECTORSOLN")
+        time_delta = timedelta(hours=12, minutes=30)
+        (
+            run_start,
+            run_end,
+            forecasted_start,
+            forecasted_end,
+        ) = self.setup_compilation_test(
+            gen_datetime, fix_forecasted_dt, forecast_type, time_delta
+        )
+        with pytest.raises(ValueError):
+            compile_raw_data(
+                run_start,
+                run_end,
+                forecasted_start,
+                forecasted_end,
+                forecast_type,
+                table,
+                raw_cache=tmp_path,
+                data_format="csv",
+            )
+
+    def test_compile_two_datetime_cols(
+        self,
+        gen_datetime,
+        fix_forecasted_dt,
+        tmp_path,
+    ):
+        (forecast_type, table) = ("STPASA", "INTERCONNECTORSOLN")
+        time_delta = timedelta(hours=12, minutes=30)
+        (
+            run_start,
+            run_end,
+            forecasted_start,
+            forecasted_end,
+        ) = self.setup_compilation_test(
+            gen_datetime, fix_forecasted_dt, forecast_type, time_delta
+        )
+        data_map = compile_raw_data(
+            run_start,
+            run_end,
+            forecasted_start,
+            forecasted_end,
+            forecast_type,
+            table,
+            raw_cache=tmp_path,
+            data_format="df",
+        )
+        runtime_col = RUNTIME_COL[forecast_type]
+        forecasted_col = FORECASTED_COL[forecast_type]
+        assert data_map is not None
+        df = data_map[table]
+        run_start = datetime.strptime(run_start, DATETIME_FORMAT)
+        run_end = datetime.strptime(run_end, DATETIME_FORMAT)
+        forecasted_start = datetime.strptime(forecasted_start, DATETIME_FORMAT)
+        forecasted_end = datetime.strptime(forecasted_end, DATETIME_FORMAT)
+        assert pd.Timestamp(df[runtime_col].unique()[0]) >= run_start
+        assert pd.Timestamp(df[runtime_col].unique()[-1]) <= run_end
+        assert pd.Timestamp(df[forecasted_col].unique()[0]) >= forecasted_start
+        assert pd.Timestamp(df[forecasted_col].unique()[-1]) <= forecasted_end
+
+    def test_compile_one_datetime_col(
+        self,
+        gen_datetime,
+        fix_forecasted_dt,
+        tmp_path,
+    ):
+        (forecast_type, table) = ("PREDISPATCH", "CASESOLUTION")
+        time_delta = timedelta(hours=2, minutes=30)
+        (
+            run_start,
+            run_end,
+            forecasted_start,
+            forecasted_end,
+        ) = self.setup_compilation_test(
+            gen_datetime, fix_forecasted_dt, forecast_type, time_delta
+        )
+        data_map = compile_raw_data(
+            run_start,
+            run_end,
+            forecasted_start,
+            forecasted_end,
+            forecast_type,
+            table,
+            raw_cache=tmp_path,
+            data_format="df",
+        )
+        runtime_col = RUNTIME_COL[forecast_type]
+        assert data_map is not None
+        df = data_map[table]
+        run_start = datetime.strptime(run_start, DATETIME_FORMAT)
+        run_end = datetime.strptime(run_end, DATETIME_FORMAT)
+        assert pd.Timestamp(df[runtime_col].unique()[0]) >= run_start
+        assert pd.Timestamp(df[runtime_col].unique()[-1]) <= run_end